requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
ultralytics>=8.2.0
pytorch-quantization
tqdm
